// InventoryManagementUI/src/openapi/services/model/inventoryRecordCreateModel.ts
/**
 * Seyir Füzeleri Sistem Mühendisliği Envanter Uygulaması API
 * NodeJS - ExpressJS API Docs / Fevzi TEMİZ
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface InventoryRecordCreateModel { 
    deliveryDate: string;
    returnDate: string;
    description: string;
    itemId: number;
    deliveredByEmployeeId: number;
    deliveredToEmployeeId: number;
}